[SearchTray] Import FormatWithIcon [TrayContainer] Import ScopeFieldsMap [SearchMetadata] Remove FormatWithIcon from component part of [#506]
