+++ src/components/TrayContainer.vue
@@ -16,6 +16,8 @@
         <SearchTray
           v-if="scope"
           :scope="scope"
+          :default-icon="scope"
+          :basic-field-list="ScopeFieldsMap[scope]"
           :results-promise="searchService.results(scope, query)"
         >
         </SearchTray>
@@ -40,6 +42,7 @@ import { TrayOrder } from '../models/TrayOrder';
 import { JumpToSectionOrder } from '../models/JumpToSectionOrder';
 import JumpToSection from './JumpToSection.vue';
 import BestBetsTray from './BestBetsTray.vue';
+import ScopeFieldsMap from '../config/ScopeFieldsMap';
 
 const query = SearchTermService.term();
 const searchService = new SearchService();
